+++ changelog.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 15.12.2022
+
+### Major changes
+
+- Add Pokemon guessing game
+
+### Changed features
+
+- Add footer top margin
+
+### Fixed bugs
+
+- Fix bug wiht correct answer dsplaying
+
 ## [0.1.0] - 15.12.2022
 
 ### Major changes